8. Added new scripts

+++ shell/git_export_src.sh
@@ -7,15 +7,24 @@ fi
 
 REPO="$1"
 
-PEPOPATH="git@your.path.com"
+read -p "clone? (Y/y/Д/д/+) " -n 1 -r
+echo ""   # (optional) move to a new line
+if [[ $REPLY =~ ^[YyДд+]$ ]]; then
+    echo "#####################################"
+    echo "#               CLONE               #"
+    echo "#####################################"
+    git clone --recursive git@git.333by.com:"$REPO"
+fi
 
-git clone --recursive $PEPOPATH:"$REPO"
+
 cd "$REPO"
 
-REV=`git rev-list --count HEAD`
-REV="${REV//[[:space:]]/}"
+# REV=`git rev-list --count HEAD`
+# REV="${REV//[[:space:]]/}"
+REV=543
 
-git log -1 --pretty=%B
+echo "Core:" > "version.txt"
+git log -1 --pretty=%B >> "version.txt"
 
 for ITEM in `git submodule`
 do
@@ -23,19 +32,40 @@ do
         cd "$ITEM"
         git checkout master
         git pull origin master
-        git log -1 --pretty=%B
+        echo "${ITEM}:" >> "../version.txt"
+        git log -1 --pretty=%B  >> "../version.txt"
         cd ..
-        rm -rf "$ITEM"/.git*
     fi
 done
 
-rm -rf .git*
+echo "#####################################"
+echo "#               CLEAR               #"
+echo "#####################################"
+
+read -p "clear? (Y/y/Д/д/+) " -n 1 -r
+echo ""   # (optional) move to a new line
+if [[ $REPLY =~ ^[YyДд+]$ ]]; then
+    # CLEAR
+    for ITEM in `git submodule`
+    do
+        if [ -d "$ITEM" ]; then
+            rm -rf "$ITEM"/.git*
+        fi
+    done
+    rm -rf .git*
+fi
+
+read -p "Pause. You can remove unnecessary files..." -n 1 -r
+echo ""
+
+echo "#######################################"
+echo "#               ARCHIVE               #"
+echo "#######################################"
 
 cd ..
 
-# read -p "Pause. You can remove unnecessary files..." -n 1 -r
-# echo ""
+zip -0 -X -r "${REPO}-src.r${REV}.zip" "${REPO}" \
+    --exclude "*.DS_Store" --exclude *.git*
 
-zip -0 -X -r "$REPO"-src\@r"$REV".zip "$REPO" \
-    --exclude "*.DS_Store"
+exit 0
